pyproject.toml
[project]
name = "kakolog"
version = "0.1.0"
description = "Long-term memory engine for Claude Code"
requires-python = ">=3.11.11"
dependencies = [
    "sentence-transformers[onnx]>=3.0",
    "sqlite-vec>=0.1.7",
    "mecab-python3>=1.0",
    "optimum>=1.0",
    "onnxruntime>=1.0",
    "sentencepiece>=0.2.1",
]

[project.scripts]
kakolog = "kakolog.cli:main"
kakolog-save = "kakolog.save:main"
kakolog-recall = "kakolog.recall:main"
kakolog-server = "kakolog.server:main"
kakolog-import = "kakolog.bulk_import:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kakolog"]
